implement hacky but working compilation
I found out that the due date for the project was earlier than I thought it was. I was forced to code this in a hurry. It works for good input, but will fail on more complicated programs – for instance, programs with multiline comments, with multiline statements, with preprocessing directives, and with spaces for indentation.